dashboard/requirements.txt
cassandra-driver>=3.25.0
pandas>=1.5.2
plotly>=5.11.0
dash>=2.7.1
